requirements.txt
uvicorn~=0.38.0
pydantic~=2.12.4
sqlalchemy~=2.0.44
alembic~=1.17.2
python-dotenv~=1.2.1
fastapi~=0.121.3
asyncpg~=0.30.0
pytest~=8.0.0
pytest-asyncio~=0.23.0
httpx~=0.27.0